Compact imported worktrees visibility prompt
